+++ examples/store-counter-react/eslint.config.js
