+++ examples/store-counter-react/src/App.tsx
@@ -1,28 +1,57 @@
 import './App.css';
 import { createStore } from '@xstate/store';
 import { useSelector } from '@xstate/store/react';
+import { useEffect } from 'react';
+import { createBrowserInspector } from '@statelyai/inspect';
 
-const store = createStore(
-  {
+const inspector = createBrowserInspector();
+
+const store = createStore({
+  context: {
     count: 0
   },
-  {
-    inc: {
-      count: (ctx) => ctx.count + 1
+  on: {
+    inc: (context, event: { by: number }) => {
+      return {
+        count: context.count + event.by
+      };
+    },
+    reset: (_context, _ev, { emit }) => {
+      emit({ type: 'reset' });
+      return {
+        count: 0
+      };
     }
   }
-);
+});
+
+store.inspect(inspector.inspect);
 
 function App() {
   const count = useSelector(store, (s) => s.context.count);
 
+  useEffect(() => {
+    const sub = store.on('reset', () => {
+      console.log('Count reset!');
+    });
+
+    return sub.unsubscribe;
+  }, []);
+
   return (
     <>
-      <h1>XState Store counter example</h1>
-      <div className="card">
-        <button onClick={() => store.send({ type: 'inc' })}>
+      <div
+        className="card"
+        style={{
+          display: 'flex',
+          flexDirection: 'row',
+          gap: '.5rem'
+        }}
+      >
+        <button onClick={() => store.send({ type: 'inc', by: 1 })}>
           count is {count}
         </button>
+        <button onClick={() => store.send({ type: 'reset' })}>reset</button>
       </div>
     </>
   );
