--- packages/xstate-store/examples/counter-react/vite.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()]
-});
